requirements.txt
slackclient
requests
Pillow
slack-sdk
